--- src/drivers/distance_sensor/sony/sony_asdt1_copy.hpp
@@ -1,132 +1,0 @@
-/****************************************************************************
- *
- *   Copyright (c) 2026 PX4 Development Team. All rights reserved.
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions
- * are met:
- *
- * 1. Redistributions of source code must retain the above copyright
- *    notice, this list of conditions and the following disclaimer.
- * 2. Redistributions in binary form must reproduce the above copyright
- *    notice, this list of conditions and the following disclaimer in
- *    the documentation and/or other materials provided with the
- *    distribution.
- * 3. Neither the name PX4 nor the names of its contributors may be
- *    used to endorse or promote products derived from this software
- *    without specific prior written permission.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
- * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
- * FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE
- * COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
- * INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
- * BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS
- * OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED
- * AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT
- * LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
- * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
- * POSSIBILITY OF SUCH DAMAGE.
- *
- ****************************************************************************/
-
-/**
- * @file sony_asdt1_copy.hpp
- *
- * Minimal AS-DT1 serial write probe.
- */
-
-#pragma once
-
-#include <px4_platform_common/module.h>
-#include <px4_platform_common/px4_config.h>
-#include <px4_platform_common/px4_work_queue/ScheduledWorkItem.hpp>
-#include <drivers/drv_hrt.h>
-#include <uORB/Publication.hpp>
-#include <uORB/topics/obstacle_distance.h>
-
-#include <mathlib/mathlib.h>
-
-#include <stdint.h>
-#include <sys/types.h>
-
-class AS_DT1_COPY : public px4::ScheduledWorkItem
-{
-public:
-	AS_DT1_COPY(const char *device, bool one_shot);
-	~AS_DT1_COPY() override;
-
-	int init();
-	void print_info();
-
-private:
-	enum class ParserState {
-		FindBegin,
-		ReadPayload,
-	};
-
-	void start();
-	void stop();
-	void Run() override;
-
-	int open_port();
-	void close_port();
-	int write_start_command();
-	int write_command_padded(const char *command);
-	int read_once();
-	bool parse_byte(uint8_t byte);
-	int process_frame(const uint8_t *frame, size_t length);
-
-	static uint32_t decode_20bit_raw(const uint8_t *data, size_t sample_index);
-	static uint16_t z_raw_to_distance_cm(uint32_t z_raw);
-	static int sample_to_layout_index(size_t sample_index);
-	static bool layout_index_to_row_col(int layout_index, int &row, int &col);
-	static int col_to_obstacle_bin(int col);
-
-	static constexpr size_t COMMAND_BUFFER_SIZE{32};
-	static constexpr size_t READ_BUFFER_SIZE{512};
-	static constexpr size_t LAST_READ_CAPTURE_SIZE{64};
-	static constexpr uint8_t BIN_COUNT = sizeof(obstacle_distance_s::distances) / sizeof(
-				obstacle_distance_s::distances[0]);
-	static constexpr size_t ASDT1_MAX_SAMPLE_COUNT{576};
-	static constexpr size_t ASDT1_BINZ_FRAME_SIZE{1440};
-	static constexpr size_t ASDT1_FRAME_BUFFER_SIZE{ASDT1_BINZ_FRAME_SIZE};
-	static constexpr int ASDT1_COLS{24};
-	static constexpr int MIN_USED_ROW{8};
-	static constexpr int MAX_USED_ROW{15};
-	static constexpr float HORIZONTAL_FOV_DEG{35.0f};
-	static constexpr float LEFT_EDGE_DEG{-HORIZONTAL_FOV_DEG / 2.0f};
-	static constexpr float OBSTACLE_INCREMENT_DEG{5.0f};
-
-	obstacle_distance_s _obstacle_distance{};
-	uORB::Publication<obstacle_distance_s> _obstacle_distance_pub{ORB_ID(obstacle_distance)};
-
-	int _fd{-1};
-	int _interval{10000};
-	char _device[20]{};
-	bool _one_shot{false};
-	unsigned int _baud{115200};
-	char _last_command[COMMAND_BUFFER_SIZE]{};
-	size_t _last_command_len{0};
-	ssize_t _last_write{-1};
-	uint64_t _read_attempts{0};
-	uint64_t _bytes_read_total{0};
-	uint64_t _no_data_reads{0};
-	uint64_t _read_errors{0};
-	hrt_abstime _last_read{0};
-	ssize_t _last_read_size{0};
-	uint8_t _last_read_bytes[LAST_READ_CAPTURE_SIZE]{};
-	size_t _last_read_bytes_len{0};
-	uint8_t _frame_buffer[ASDT1_FRAME_BUFFER_SIZE]{};
-	size_t _frame_buffer_len{0};
-	ParserState _parser_state{ParserState::FindBegin};
-	size_t _begin_match_index{0};
-	uint64_t _frames_rx{0};
-	uint64_t _frames_pub{0};
-	uint64_t _parser_resets{0};
-	size_t _last_frame_processed_len{0};
-	size_t _last_sample_count{0};
-	uint8_t _last_valid_bins{0};
-	uint16_t _last_closest_distance{UINT16_MAX};
-};
